backend/requirements.txt
distlib==0.3.9
filelock==3.17.0
platformdirs==4.3.6
virtualenv==20.29.2
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
pika==1.3.2
drf-yasg==1.21.10
